Update patch set 5
Patch Set 5:

(2 comments)

> Patch Set 3: Code-Review+1
> 
> (2 comments)
> 
> Sorry to ask stupid question. I understand the purpose of this commit. However, do we need the interface to get the resource from framework? I mean that we would like the updatability of NetworkStack but we will leave some configurations in framework. The configuration will not able to be updated with NetworkStack. 
> 
> May I know what is the idea here? Do we still like to other OEM manufacturers to use their own customization or configuration? e.g. config_captive_portal_http_url? OEM(or us) could put their own overlay to customize it.
> 
> Do we have any resource that may be updated along with NetworkStack?

Good point, I added more justification to the commit message. There are 2 cases:
 - Device uses the network stack module. Then customization must be done with a resource overlay, and default values are defined in strings.xml in the app (code will use fallback values if strings are null/empty)
 - Device does not use the network stack module, and runs the network stack in-process in the system server. Customization must be done by adding values to the framework strings.xml.

The objective is mainly to reduce code edits in the network stack, since it's not possible in the first case, and it would make it more difficult for OEMs to migrate from the second case to the first one.

The overlay values would not be updated with the network stack, although OEMs might be able to update their overlay package if they use the updatable network stack.

Patch-set: 5
Reviewer: Gerrit User 1288034 <1288034@85c56323-6fa9-3386-8a01-6480fb634889>