Update patch set 5
Patch Set 5:

(6 comments)

I'm not quite sure how to use this class because I couldn't find a concrete example (the example resource here is used in NetworkMonitor, but I don't see how NetworkMonitor actually uses the ResourceManager to find the resource).

But if what i gather from the CL is correct, then this class seems a bit harder to use than I would have hoped, since users have to migrate from r.getInteger(...) to ResourceManager.getXyz().

Generally the usage of resources is:

    Resources r = mContext.getResources();
    int fooBar = r.getInteger(R.integer.config_fooBar);

The first doesn't seem too hard to make generic, since the actual resources object to be used is always an instance of Resources, and thus can be returned by a method that either calls mContext.getResources(), which returns app resources, or mContext.getResources().getSystem(), which returns system resources.

As for the second one, the difficulty is making sure that an in-process network stack gets com.android.internal.R, and the app-based networkstack gets... something else? Is this something that can be solved in a different way than you're doing here? For example, is it possible to use jarjar to rename com.android.internal.R in the app?

Patch-set: 5
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>