Update patch set 3
Patch Set 3:

This won't merge as is into AOSP.

CONFLICT (content): Merge conflict in core/jni/AndroidRuntime.cpp
CONFLICT (content): Merge conflict in core/jni/Android.bp

Not sure what to do about this. If there's a hacky way to get this to merge in AOSP by reordering the lines in AOSP first, and then applying this change, then we might want to try that. The advantage of doing this is that the change only has to be reviewed once.

If that doesn't work, we'll have to merge this into master first and backport it with Merged-In.

Patch-set: 3
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>