Protect against unknown parcelables
https://issuetracker.google.com/issues/128507733

The problem is that ResolverActivity (via ResolverComparator)
attempts to unparcel the target intent's extras to pull out
the content annotations extras. If there are any classes
in there that it is unaware (such as any custom Parcelable in
the app) we will crash.

Make a copy of the original target intent before we attempt to
unparcel it, to ensure we do not corrupt it after a BadParcelException.
Catch BadParcelException when trying to pull out the content
annotations and fail gracefully.

Change-Id: I645e965ff22bf7d4532526d1e9a9569ac836a1c1
Signed-off-by: Jeffrey Blattman <jeffrey.blattman@gmail.com>