Preload non-bootclasspath Java libraries.
This is specifically for HIDL but is applicable to other libs.

Classes on the bootclasspath are implicitly used by apps. For this
reason, many classes should not go there. However, there are some
classes which are used by many apps/processes which are still nice
to preload.

Now, cacheNonBootclasspathSystemLibs in ApplciationLoaders keeps
a map of jar -> ClassLoader in zygote to be retrieved by child
processes.

Bug: 128529256
Bug: 127406460
Test: boot Pixel 2, verify libs are preloaded and used, try apps that
    use these libraries.
Test: grep for ClassLoaderContext errors, for instance:
 - ClassLoaderContext shared library size mismatch
 - ClassLoaderContext classpath element mismatch
Except for unrelated error:
03-22 03:14:02.802 20091 20091 W system_server: ClassLoaderContext classpath size mismatch. expected=3, found=0 (PCL[services.jar*4172153037:ethernet-service.jar*2871476484:wifi-service.jar*2460929558] | PCL[])
Test: showmap on various processes which use the preloaded libs.

Change-Id: I351bf1679e9a928c10dca860b6cd6cb414c3bb8e