Preload non-bootclasspath Java libraries.
This is specifically for HIDL but is applicable to other libs.

Classes on the bootclasspath are implicitly used by apps. For this
reason, many classes should not go there. However, there are some
classes which are used by many apps/processes which are still nice
to preload.

Now, cacheNonBootclasspathSystemLibs in ApplciationLoaders keeps
a map of jar -> ClassLoader in zygote to be retrieved by child
processes.

Bug: 128529256
Bug: 127406460
Test: wip
Change-Id: I351bf1679e9a928c10dca860b6cd6cb414c3bb8e