Deprecate getCaptivePortalServerUrl
This API cannot return correct values as network validation is now
managed by the updatable NetworkStack, which may use configurable URLs,
changing URLs, or mechanisms not involving URLs to detect captive
portals.

With this deprecation, the source of truth for current URLs is moved
from the system server to NetworkStack resources. A legacy default value
is left in ConnectivityService for the deprecated API.

Test: flashed, booted, WiFi and captive portal working
Bug: 127908503
Change-Id: I371dedc5b22efa909d7fd58e1ebe9b8aaced9780