Update patch set 21
Patch Set 21:

(12 comments)

How does this interact with 464xlat ? IIUC this is called in NetworkStatsRecorder, while 464xlat adjustments are done in NetworkStatsService, after these tun adjustments have been made.

The VPN declares underlying networks, whose interfaces are the physical interfaces. But IIUC if the VPN is handling any v4 traffic that amount will be reflected both on the v4- interface and the physical one. This is going to subtract from the stats billed to the VPN the part of traffic on the physical interface that should not be billed to it, but not touch the stats on the v4- interface. Then 464xlat adjustments will come and add per-packet header sizes to apps.

If I've got this right, the totals still should add up after tun adjustment, modulo rounding errors. Because everything is always rounded down, it's possible that there would be a few missing bytes that won't be subtracted from uid 0 but not that these end up with negative values, so that's probably okay... assuming I'm not getting anything backward, because if I am we'll see negative counters popping up again.

However the packet counts are estimates (and for good reason – the app didn't actually do this, it's scaled values to be representative). If app A does only v4 and app B does only v6 and assuming the VPN simply writes these as is, app B gets billed for a part of app A's computed IPv6 header overhead ? I don't think this is bad (it's probably mostly unavoidable, and within reasonable error margins) unless it may cause some stats to roll back, which a this point I can't predict.

Junyu, you know this part much better than I do, do you have any thoughts ?

Patch-set: 21
Reviewer: Gerrit User 1015428 <1015428@85c56323-6fa9-3386-8a01-6480fb634889>