Update patch set 23
Patch Set 23:

(1 comment)

> Hmm....that is a great question. I'm not sure what the correct answer is tbh. From what I can tell, tunMigration never has taken stacked interfaces into account (and this CL has not modified that behavior in any way). Now it is totally possible that the existing behavior is possibly broken, and maybe we should fix it. I'm going to leave that for you to decide since you are the expert in 464xlat stuff (and I don't know how it works in detail). The change however should be trivial if we do think we should also add stacked ifaces (NetworkStatsFactory#augmentWithStackedInterfaces already does this).
> 
> Another thing to consider is what's the side affect of *not* including stacked ifaces. AFAICT, traffic for stacked ifaces (from VPN uid) would not get distributed to apps. The question is does that matter? Does the data accounting cares about stacked ifaces or just underlying ifaces?

IIUC, yes it is counted. The mActiveUidIfaces takes the stacked iface at below line.
https://cs.corp.google.com/aosp-master/frameworks/base/services/core/java/com/android/server/net/NetworkStatsService.java?q=NetworkStatsService&sq=package:aosp-master&dr=CSs&l=1221

Anyway if it is a real bug, then it is already broken long time ago.
In any case it doesn't have to be fixed in this change.
Here is how to verify it:
  1.connect to an ipv6-only wifi
  2.connect to an vpn that only works on ipv4.
  3.download 10mb file by using browser
  4.go to detail usage page and verify if the usage for browser is increased around 10mb.
I think we should at least memo it somewhere.
If you are not feasible to verify this, could you help to open a bug and assign to me? I'll find sometime to do it.

Patch-set: 23
Reviewer: Gerrit User 1306837 <1306837@85c56323-6fa9-3386-8a01-6480fb634889>