Update patch set 1
Patch Set 1: Code-Review-1

The change itself is benign enough, but it is completely broken that it's needed.

This is conflicting because the IKE test build target, which builds a package with a manifest, is including the NetworkStackBase target, which is a package with a manifest.
Worse, the regular IKE target, which is a java library, is including the NetworkStackBase target with its manifest. Any package with a label that wants to use the IKE library will experience the same build breakage.

I did some digging to figure out why IKE is using this target. It turns out it wanted to use PacketReader, which is a small utility which happens to live in this package because it's where it used to lived.

Instead of doing this, please only include what you need instead of the entire NetworkStackBase package inside IKE. It should be easy enough to create a filegroup that contains just PacketReader, or better – move PacketReader to some common area and make a static lib that can be used by the network stack, IKE, and possibly other clients.

Patch-set: 1
Reviewer: Gerrit User 1015428 <1015428@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1