Fix repetitiv key in BluetoothEventManager.mHandlerMap
In the constructor of BluetoothEventManager, it put BluetoothHeadset.ACTION_AUDIO_STATE_CHANGED to mHandlerMap.

And then, in the method updateLocalProfiles of LocalBluetoothProfileManager,
it put BluetoothHeadset.ACTION_AUDIO_STATE_CHANGED to mHandlerMap again.

As result, the AudioModeChangedHandler has been replaced by the HeadsetStateChangeHandler.

So it need a independent Map for mProfileBroadcastReceiver.