Fix repetitiv key in BluetoothEventManager.mHandlerMap
In the constructor of BluetoothEventManager, it put BluetoothHeadset.ACTION_AUDIO_STATE_CHANGED to mHandlerMap.

And then, in the method updateLocalProfiles of LocalBluetoothProfileManager,
it put BluetoothHeadset.ACTION_AUDIO_STATE_CHANGED to mHandlerMap again.

As result, the AudioModeChangedHandler has been replaced by the HeadsetStateChangeHandler.

So it need a independent Map for mProfileBroadcastReceiver.

Fix:129650132

Test:manual
1. Connect a bluetooth headset.
2. Receive a incoming call.
3. Hangup the call, check the sub-title in "Connected devices" page.

Change-Id: I3ed3764c8a00d824c768ae7ca53948dd66eadd8e