Update patch set 4
Patch Set 4: Code-Review-2

After discussion with Suren, the function call is currently in the wrong place and will result in it either being called on every app launch or never called, depending on if the USAP pool is enabled or not.  The correct place to put this doesn't currently exist, so I'll be adding it today and then we can re-base this CL onto it.

Patch-set: 4
Reviewer: Gerrit User 1047255 <1047255@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2