Update patch set 2
Patch Set 2:

Who is the right person to review this CL and the base CL for +2 purposes?

I don't want to step on toes and this approach is not what I was originally hoping for (the fact that there are three calling threads from two processes makes it hard to avoid potentially blocking for locks which I'd have liked to avoid), but I've also not seen any other CL that convinced me personally as being cleaner / better. 

You're the owners, it's completely up to you whether you want to +2 this CL or try a different approach, but I think we need *some* fix soon (we've already taken longer than I would prefer).

Note that if we want to submit this CL, then I'd need +2 both on this CL and on the base CL.

Patch-set: 2
Reviewer: Gerrit User 1001525 <1001525@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1103203 <1103203@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0