Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Yan, could you let me know if you have or are planning on doing this soon? If not then I might have time to look at it.
> 
> Hi Paul, we are going to make a copy of PacketReader in ike and do not include NetworkStackBase anymore. There is a CL for that under review. I'll let you know when we merge that CL.

Just a flavor note, nobody is happy about making a copy of these files, but bikeshedding on an alternative is imprudent right now, so we're keeping a bug and a TODO to see if we can achieve consensus on a better approach. A quick chat with Connectivity folks did not immediately generate such a consensus (hence, kicking the can down the road).

Patch-set: 1