Update patch set 2
Patch Set 2:

Actually we could even make ourselves a utility that does exactly this : use the async API in a sync way. It would be useful in cases like this where the async API supports stuff that the sync API does not. We'll be in a better world when we can use coroutines, and NetworkMonitor is the prime candidate for conversion in that it would simplify it to an extreme degree.

Patch-set: 2
Reviewer: Gerrit User 1015428 <1015428@85c56323-6fa9-3386-8a01-6480fb634889>