Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > Could we depend on it as a library, rather than including the source directly?
> > 
> > I don't know if that'd make much of a difference? These are the exceptions, where explicit dependency on the annotation is needed, otherwise it gets picked up transitively.
> 
> Hmm, I'm not really familiar enough with the build. It sounds like you need to *export* the annotation from these targets, rather than just depend on it?
> 
> Including the source like this means it gets compiled multiple times, I don't know how much that matters.
> 
> Maybe David knows more?

High level:
I know this is a cleanup because technically we only need one annotation at the moment, but be careful. We are likely to start building libcore and framework independently of each other in the future. The annotation is internal-only and build-only, so having a dependency from the framework would require having libcore export it and keep it in the binary module. I would stick to two annotation.

Low level:
Don't put the same class in the boot classpath twice. There should be a build-time assertion against it. Best to build against a library.

Patch-set: 2
Reviewer: Gerrit User 1059815 <1059815@85c56323-6fa9-3386-8a01-6480fb634889>