Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 2:
> > > > 
> > > > Could we depend on it as a library, rather than including the source directly?
> > > 
> > > I don't know if that'd make much of a difference? These are the exceptions, where explicit dependency on the annotation is needed, otherwise it gets picked up transitively.
> > 
> > Hmm, I'm not really familiar enough with the build. It sounds like you need to *export* the annotation from these targets, rather than just depend on it?
> > 
> > Including the source like this means it gets compiled multiple times, I don't know how much that matters.
> > 
> > Maybe David knows more?
> 
> High level:
> I know this is a cleanup because technically we only need one annotation at the moment, but be careful. We are likely to start building libcore and framework independently of each other in the future. The annotation is internal-only and build-only, so having a dependency from the framework would require having libcore export it and keep it in the binary module. I would stick to two annotation.
> 
> Low level:
> Don't put the same class in the boot classpath twice. There should be a build-time assertion against it. Best to build against a library.

Hmm, I would rather have a single annotation if we can; with two we just risk them getting out of sync, and complicates other things.

Since it's build-only, could we create a new target that can be depended on by both libcore and frameworks? Logically, it shouldn't need to be in any binary module, since it's not needed at runtime (though I'm not sure exactly what's meant by a binary module).

(maybe the solution is to create a third annotation in a new place, and migrate both the existing ones to that?)

Patch-set: 2
Reviewer: Gerrit User 1010681 <1010681@85c56323-6fa9-3386-8a01-6480fb634889>