No longer show disambig when calling SmsManager#getSubId
SmsManager#getSubId used to show a disambig dialog to
pick a default SMS subscription when
SmsManager#getSubscriptionId was called. Since it was used
internally as well, it would cause many of these dialogs
to pop up over and over again.

Do not display the dialog anymore and instead move the
dialog pop up to the methods that actually require the use
of the dialog, such as sending a text.

Bug: 130853716
Test: Manual, Default SMS app
Merged-In: I71ad38f76750c750b149e4cdcb6fcf99ed89212e
Change-Id: I71ad38f76750c750b149e4cdcb6fcf99ed89212e