Update patch set 2
Patch Set 2:

--

Eric,

CS = Class Specific . Video and Audio descriptors have the same id 0x24 (CS_INTEFACE) . The current code is not correct, because it assumes

CONFIDENTIALITY NOTICE: The contents of this email message and any attachments are intended solely for the addressee(s) and may contain confidential and/or privileged information and may be legally protected from disclosure. If you are not the intended recipient of this message or their agent, or if this message has been addressed to you in error, please immediately alert the sender by reply email and then delete this message and any attachments. If you are not the intended recipient, you are hereby notified that any use, dissemination, copying, or storage of this message or its attachments is strictly prohibited.

Patch-set: 2
Tag: mailMessageId=<CAGRzWkrCoDzqxtoMk=MTYux=uV-5Avq2w_=Oas4kP-c+JYddew@mail.gmail.com>