Update patch set 1
Patch Set 1:

I don't really like this approach.

Note that it looks important to be able to compile sometimes, e.g., to avoid ANRs in mega-apps.

It seems better to "just" avoid the path we know is bad - the BackgroundDexoptService, and leave full compile open for the user to manually trigger.

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>