Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1:
> 
> I don't really like this approach.
> 
> Note that it looks important to be able to compile sometimes, e.g., to avoid ANRs in mega-apps.
> 
> It seems better to "just" avoid the path we know is bad - the BackgroundDexoptService, and leave full compile open for the user to manually trigger.

I decided to stick with this version for now. Yes, you are right that this is tied to a particular runtime behaviour and if that changes it will be obsolete. The loose coupling is indeed unfortunate. However, given that there's no clear better way to "fix" all cases (e.g. a/b; the fact we might still wanna quicken them and not complete ignore the packages) I think it's better to do it in PDO. We already modify the compilation filter for other cases there too, and heaving it all filter-change related logic in one place is at least slightly better.

Patch-set: 2
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>