Net: Register IpConnectivityMetrics service in onStart
ConnectivityService unconditionally expects IpConnectivityMetrics logger
service to be always registered to LocalServices, and especially expects
to be able to log metrics events when a NetworkAgent is registered.

However IpConnectivityMetrics only registers its logger service to
LocalServices in the PHASE_SYSTEM_SERVICE_READY boot step, which can
cause crashes in the system server if a network service tries to
register a NetworkAgent as soon as it receives the signal for
PHASE_SYSTEM_SERVICE_READY.

To solve this issue this patch moves IpConnectivityMetrics logger
service registration earlier in the boot sequence in onStart(), where it
would be incorrect for a network bearer to register a NetworkAgent to
ConnectivityService.

In addition, NetdEventListenerService now call Context#getSystemService
again if its current ConnectivityManager handle was null. Previously it
was safe for NetdEventListenerService to assume it would never get a
null ConnectivityManager reference when starting in
PHASE_SYSTEM_SERVICE_READY, but that does not hold anymore if its parent
IpConnectivityMetrics service now creates it in onStart().

Bug: 132766803
Bug: 133367806
Test: Compiled, flashed crosshatch, booted, checked dumpsys connmetrics.
Change-Id: I07e218533f5c99bd71f85f775200004147ada91d