Update patch set 1
Patch Set 1: Code-Review-1

Thanks for reaching out. I think we have a similar fix for this already in Q. We're also looking to rewrite this class, hopefully for R, to make it more testable and easier to understand (you've probably noticed it could be improved in that way :-)

So I'm not inclined to take the patch, since it won't merge cleanly with the changes underway.

Patch-set: 1
Reviewer: Gerrit User 1092253 <1092253@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1