Update patch set 3
Patch Set 3: Code-Review-1

Hi Sanshan,

Thanks for the contribution. I'm not positive on this CL as only adding a time-out doesn't solve the underlying issue, just makes it disappear in some circumstances. 

The correct fix would be to wait on the agent to return before resuming the operation. We've already implemented this fix in Q, so this won't be a problem in CTS for Q+. We can't backport the fix to P because it requires new infra code that we added in Q, so it's not a safe fix at this point.

+Ross Yeah for CTS

Ross, 

We have a CTS test (2 actually: KeyValueQuotaTest#testQuotaExceeded, FullBackupQuotaTest#testQuotaExceeded) that found an actual error in the code-under-test. We've hit this before and landed the fix in the code-under-test in Q. So, CTS should pass for Q+ devices but is flaky on P. There is no workaround for this on the test side, should we disable the test for P?

Patch-set: 3
Reviewer: Gerrit User 1431448 <1431448@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1244622 <1244622@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1133744 <1133744@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1