Update patch set 3
Patch Set 3:

Hey Sanshan,

The problem in putting a number like 20 ms is that it won't fix the problem on all devices due to different configurations among them and hence different times for execution. Even though it might work multiple times on the same platform it may not work under different scenarios that you are not simulating. On top of that agent calls are one-way, which means there is no guaranteed execution of the quota exceeded call when it returns, it will be executed on a binder thread pool in the other process, without guaranteed time of execution either. I'm strongly against having such timers in the code as it makes the code extremely brittle and very hard to reason about.

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3: Code-Review-1
> > > 
> > > Hi Sanshan,
> > > 
> > > Thanks for the contribution. I'm not positive on this CL as only adding a time-out doesn't solve the underlying issue, just makes it disappear in some circumstances. 
> > > 
> > > The correct fix would be to wait on the agent to return before resuming the operation. We've already implemented this fix in Q, so this won't be a problem in CTS for Q+. We can't backport the fix to P because it requires new infra code that we added in Q, so it's not a safe fix at this point.
> > > 
> > > +Ross Yeah for CTS
> > > 
> > > Ross, 
> > > 
> > > We have a CTS test (2 actually: KeyValueQuotaTest#testQuotaExceeded, FullBackupQuotaTest#testQuotaExceeded) that found an actual error in the code-under-test. We've hit this before and landed the fix in the code-under-test in Q. So, CTS should pass for Q+ devices but is flaky on P. There is no workaround for this on the test side, should we disable the test for P?
> > 
> > If there is not workaround, please add the tests to known failure list to skip them. Please also create a bug for our tracking as well.
> 
> Hi Bernardo&Ross,
> we can just add 20ms which is enough for thread mBackupRunner to execute related function. And it will not impact normal backup function, it will only impact the exceeded quota case.
> It can get positive result if add some delay here in all our imx devices.

Patch-set: 3
Reviewer: Gerrit User 1133744 <1133744@85c56323-6fa9-3386-8a01-6480fb634889>