Update patch set 2
Patch Set 2: Code-Review+1

I have no +2 vote unfortunately.
AFAICT this will be automerged to the internal master, so no need to cherry-pick.
On the other hand, if this is important to go in Q, it should be cherry-picked to qt-dev.

Patch-set: 2
Reviewer: Gerrit User 1292888 <1292888@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1