Update patch set 4
Patch Set 4:

Agreed that this is a problem, but I'm not sure the fix is sane here. The paddings that are selected are absolute, such that they align with where the icon and the text normally is when it is not heads upped. Those contents are not inset by the padding provided, which would then be visible. Can we make sure we're not clipping it instead?

Patch-set: 4
Reviewer: Gerrit User 1060062 <1060062@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1059880 <1059880@85c56323-6fa9-3386-8a01-6480fb634889>