Update patch set 4
Patch Set 4:

hi Peter

As I remembered, the left boundary of the icon on status bar aligns with the title of the notification for UI/UX experience.

Yes, as your mention. The rounded_corner_content_padding doesn't be considered but ...
When increasing the rounded corner radius, the vertical notification position should/would change with the corner radius. Without changing, the notification boundary would be clipped by the round corner. And, it would make the top padding of the notification panel increase and the padding is too much. It's a trade-off.

Maybe, the requirement is to make the accessibility better so decrease the density and then the everything increase the size. of course, the size of the corner radius increase. But, it should change a little but not a lot. The major point is to increase the content/context size rather the corner radius size.

To increase the corner radius should also increase the notification_content_margin_start. Each vendor could have itself design and look and feel.

Patch-set: 4
Reviewer: Gerrit User 1309980 <1309980@85c56323-6fa9-3386-8a01-6480fb634889>