Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> hi Peter
> 
> As I remembered, the left boundary of the icon on status bar aligns with the title of the notification for UI/UX experience.
> 
> Yes, as your mention. The rounded_corner_content_padding doesn't be considered but ...
> When increasing the rounded corner radius, the vertical notification position should/would change with the corner radius. Without changing, the notification boundary would be clipped by the round corner. And, it would make the top padding of the notification panel increase and the padding is too much. It's a trade-off.
> 
> Maybe, the requirement is to make the accessibility better so decrease the density and then the everything increase the size. of course, the size of the corner radius increase. But, it should change a little but not a lot. The major point is to increase the content/context size rather the corner radius size.
> 
> To increase the corner radius should also increase the notification_content_margin_start. Each vendor could have itself design and look and feel.

Well, though I don't think the rounded corner of the notification cards should change with the rounded corner of the screen itself, whether these values should change together seems more of a design problem now (I personally don't like a much larger rounded corner on notification cards though). But if they should be correlated, I think they should be somehow entangled in the code rather than being separate attributes.

Anyhow, what I am sure here is that on some devices, it is okay to keep the rounded corner of the notification cards intact while having round corners, because they don't conflict with each other. But the heads-up icon do get clipped, which isn't exactly the same thing as a notification card, and isn't something I'd expect to happen. Plus that it seems that on a device with corner cutout, from the code, the heads-up icon and text are moved away from the cutout regardless of the alignment with notification cards, which is kind of a similar case to when a device have a huge rounded corner.

I agree the current patch isn't very sane and doesn't look very good either. Maybe for now let's just ensure the icon is not clipped in any case, either by capping the padding to 0 or by allowing children to draw outside of the parent view? (I'd prefer capping the padding to 0 though, because it at least aligns with the clock position, which is better than being too close or swallowed by the rounded corner)

Patch-set: 4
Reviewer: Gerrit User 1500189 <1500189@85c56323-6fa9-3386-8a01-6480fb634889>