Update patch set 4
Patch Set 4:

If you have a fix to fix the clipping by allowing it to draw outside the boundary, I'd be glad to take that. Usually that should be fine if the overlap is just a little bit.
Agreed that if the overlap is a LOT it probably makes sense to take a different approach, like insetting it a bit more. That's very device dependent though, so I'd rather not include that.

Patch-set: 4
Reviewer: Gerrit User 1059880 <1059880@85c56323-6fa9-3386-8a01-6480fb634889>