Add DENSITY_450, DENSITY_200
Add 450dpi and 200 dpi as a supported screen density.
This change follows up below change Id : Iac673eae2d82c27d96e279c076454af2f5933208

Originally 600dpi was added to support sw384 with WQHD resolution,
though it was lack of for the cases when resolution is FHD and WVGA resolution. (In case of HD, there already exists 300 for HD resolution).

Test: run android.dpi.cts.ConfigurationTest#testScreenConfiguration and android.app.cts.ActivityManagerMemoryClassTest#testGetMemoryClass

Change-Id: I01d5d46d5cbbf79dd30540b11d7aa6906938a662
Signed-off-by: bj1211jeon <bj1211.jeon@samsung.com>