Add host-driven tests for the NetworkStack
Add a first test to verify that the NetworkStack can be installed and
rolled back.
The test Will only pass on a device built with the same key
configuration as the test. It will not pass on production devices for
example.
The test also uses the NetworkStack in the tree, so it does not test
mismatches between the version of the NetworkStack and that of the
system.

On devices where adb root is available, the test performs the following:
 - Install the NetworkStack APK with --staged and --enable-rollback
 - Reboot the framework
 - Check that the network stack is registered
 - Request a rollback of the network stack
 - Reboot the framework
 - Check that the network stack is registered again

When adb root is not available, the test performs a reboot instead of a
framework restart. The test then takes 2m25s instead of 25s.

Test: atest NetworkStackHostTests
Bug: 133025642
Change-Id: I4d97db4b4b5052d99b986560cc52c3f031c6907a