Update patch set 1
Patch Set 1:

> Hi Michael,
> 
> Is it going to make any difference if the new WindowManagerPolicy is extending WindowManagerPolicy (the interface) instead of PhoneWindowManager (the phone implementation)? WindowManagerPolicy seems to be a stable (stabler?) API surface than PWM.
> 
> And maybe we should mark PhoneWindowManager as final if we do not consider it an API? It seems like there is only one class (PhoneWindowManagerArc) that extends it.


Hi Haoxiang,

Just to add my opinion, 

Extending directly the WindowManagerPolicy (the interface) instead of PhoneWindowManager (the phone implementation) is not feasible...

It is because the PhoneWindowManager has too much core implementation of how "Window" works,
we cannot bypass the implementations of PhoneWindowManager when creating a new WindowManagerPolicy,
(It seems, the PhoneWindowManager was named as "Phone"WindowManager, when AOSP was only for Phone a long time ago...?)

Therefore, when creating a new WindowManagerPolicy, it should extend the PhoneWindowManager.
(meaning, the PhoneWindowManager should not be changed as final...)

That's my thought. :)

-------------------------------------------------------------------------------------

To everyone,

In order to have AOSP compatible with many devices(not just phone and car)/vendors/countries/regulations, it should be able to allow adding vendor's own window types in some way (official way)..

Patch-set: 1
Reviewer: Gerrit User 1463516 <1463516@85c56323-6fa9-3386-8a01-6480fb634889>