Don't merge interlaced and non-interlaced modes together
When a display supports a resolution in interlaced and progressive, these
modes will be merged together in the local display adapter since it only
compares width/height/refresh. Depending on the order of the modes
provided by HWC, the interlaced mode may be chosen instead of the more
desirable progressive mode.

This patch adds a new interlaced flag to Display.Mode so we can
differentiate between interlaced and progressive modes when looking for
the default mode.

Ideally we'd have a lot more information (like subsampling, preferred
bit, porch info, etc), but this is at least a conversation starter.

Test: Tested on hikey960 and externally on Amlogic G12A

Reported-by: Neil Armstrong <narmstrong@baylibre.com>
Tested-by: Neil Armstrong <narmstrong@baylibre.com>
Signed-off-by: Sean Paul <seanpaul@chromium.org>

Change-Id: Iaeea58834aa19c352cc56d19ba5e40eb39374e4f