Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hi, I'm curious what the motivation is for proposing this change be upstreamed to AOSP?  The changes proposed are for a niche set of functionality which do not seem applicable to the ecosystem as a whole.

It'll make it easier for third party developers of push to talk applications to have a common key code among different devices with physical PTT buttons (of which Yury/Motorola listed a few, but there are more). Instead of the developer having to handle different vendors in a different way.

Furthermore, it says in KeyEvent.java that:

    //  Also Android currently does not reserve code ranges for vendor-
    //  specific key codes.  If you have new key codes to have, you
    //  MUST contribute a patch to the open source project to define
    //  those new codes.  This is intended to maintain a consistent
    //  set of key code definitions across all Android devices.

Even if Android had a key range for vendor keys, it would require all vendors to agree on a common keycode. It would be beneficial for Android, manufacturers and developers, that one could agree on a few key codes for these use cases that are very important for increasing Androids marketshare among enterprises. The goal of PTT is to replace radios with smartphones (increasing Androids marketshare).

Patch-set: 2
CC: Gerrit User 1065940 <1065940@85c56323-6fa9-3386-8a01-6480fb634889>