Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (11 comments)
> 
> > Even if Android had a key range for vendor keys, it would require all vendors to agree on a common keycode. It would be beneficial for Android, manufacturers and developers, that one could agree on a few key codes for these use cases that are very important for increasing Androids marketshare among enterprises. The goal of PTT is to replace radios with smartphones (increasing Androids marketshare).
> 
> 
> This is, in fact, exactly why Android does not have a range of vendor keys.

The need here is for a generic intent naming defined for OEMs to use to enable the ecosystem of Solution Partners to use.  We support the submission from Moto but are looking for an agreed name that can be used now.  Even if the Android Team wanted to be even more generic PB1 and PB2 (for Programmable Button 1 and Programmable Button 2) we would be fine.  Currently the LOE is too great due to the lack of standardization.  

As for the behavior that has been mentioned.  We feel that it should far under the same path as assigning a default app (like messenger, browser, etc).  
  - If the user has 2 or more applications that are listening for the intent when the button is pressed then the user would be presented with a list app applications that can use the intent.  
  - The user would select Always to tie the app function (vs. opening the app) to this particular button press. 
  -If/When the user would like to remap this button to a different app's function then they would go into the settings menu (possibility even in the existing Default Applications menu) and remove or select a different app).
      --- this menu would only appear in the event the OEM has placed buttons with this capability... otherwise this meu would be hidden.

There is a long list of applications and OEM devices that this can be used.  This would open a large and completive landscape that is currently being underserved and address in a very fragmented way.

Patch-set: 2
CC: Gerrit User 1502259 <1502259@85c56323-6fa9-3386-8a01-6480fb634889>